Fixes issue #2308 - Rename policy extension to standard-policy
