+++ extension/standard-policy/src/main/java/cloud/piranha/extension/standard/policy/StandardPolicyExtension.java
@@ -25,36 +25,45 @@
  * ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
  * POSSIBILITY OF SUCH DAMAGE.
  */
-package cloud.piranha.extension.policy;
+package cloud.piranha.extension.standard.policy;
 
+import cloud.piranha.core.api.WebApplication;
+import cloud.piranha.core.api.WebApplicationExtension;
+import static java.lang.System.Logger.Level.DEBUG;
+import static java.lang.System.Logger.Level.WARNING;
+import java.security.NoSuchAlgorithmException;
 import java.security.Policy;
-import java.util.HashMap;
 
 /**
- * The class to keep track of setting/removing the thread local for Policy.
+ * The WebApplicationExtension that is responsible for setting up the proper
+ * Policy instance so it can be made available during web application
+ * initialization and subsequently during request processing as well as
+ * delivering listeners to set/remove the Policy from the current thread.
  *
  * @author Manfred Riem (mriem@manorrock.com)
  */
-public class PolicyThreadLocal {
+public class StandardPolicyExtension implements WebApplicationExtension {
 
     /**
-     * Stores the policies by thread id.
+     * Stores the logger.
      */
-    private static final HashMap<Long, Policy> POLICIES = new HashMap<>(1);
+    private static final System.Logger LOGGER = System.getLogger(StandardPolicyExtension.class.getName());
 
     /**
-     * Remove the policy.
-     */
-    public static void removePolicy() {
-        POLICIES.remove(Thread.currentThread().getId());
-    }
-
-    /**
-     * Set the policy.
+     * Configure the web application.
      *
-     * @param policy the policy.
+     * @param webApplication the web application.
      */
-    public static void setPolicy(Policy policy) {
-        POLICIES.put(Thread.currentThread().getId(), policy);
+    @Override
+    public void configure(WebApplication webApplication) {
+        try {
+            LOGGER.log(DEBUG, "Configuring webapplication");
+            Policy policy = Policy.getInstance("JavaPolicy", null);
+            webApplication.setAttribute(Policy.class.getName(), policy);
+            StandardPolicyThreadLocal.setPolicy(policy);
+            webApplication.addListener(StandardPolicyServletContextListener.class.getName());
+        } catch (NoSuchAlgorithmException ex) {
+            LOGGER.log(WARNING, "Error setting up Policy", ex);
+        }
     }
 }
