+++ extension/standard-policy/src/main/java/cloud/piranha/extension/standard/policy/StandardPolicyServletContextListener.java
@@ -25,39 +25,28 @@
  * ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
  * POSSIBILITY OF SUCH DAMAGE.
  */
-package cloud.piranha.extension.policy;
+package cloud.piranha.extension.standard.policy;
 
-import cloud.piranha.core.api.WebApplication;
-import jakarta.servlet.ServletRequestEvent;
-import jakarta.servlet.ServletRequestListener;
+import jakarta.servlet.ServletContextEvent;
+import jakarta.servlet.ServletContextListener;
 import static java.lang.System.Logger.Level.DEBUG;
-import java.security.Policy;
 
 /**
- * The ServletRequestListener that sets the Policy instance on the current
- * thread just before request processing and removes the Policy instance from
- * the current after the request has been processed.
+ * The ServletContextListener used to remove the Policy instance once
+ * initialization is done.
  *
  * @author Manfred Riem (mriem@manorrock.com)
  */
-public class PolicyServletRequestListener implements ServletRequestListener {
+public class StandardPolicyServletContextListener implements ServletContextListener {
 
     /**
      * Stores the logger.
      */
-    private static final System.Logger LOGGER = System.getLogger(PolicyServletRequestListener.class.getName());
+    private static final System.Logger LOGGER = System.getLogger(StandardPolicyServletContextListener.class.getName());
 
     @Override
-    public void requestDestroyed(ServletRequestEvent event) {
+    public void contextInitialized(ServletContextEvent event) {
         LOGGER.log(DEBUG, "Removing Policy");
-        PolicyThreadLocal.removePolicy();
-    }
-
-    @Override
-    public void requestInitialized(ServletRequestEvent event) {
-        LOGGER.log(DEBUG, "Setting Policy");
-        WebApplication webApplication = (WebApplication) event.getServletContext();
-        Policy policy = (Policy) webApplication.getAttribute(Policy.class.getName());
-        PolicyThreadLocal.setPolicy(policy);
+        StandardPolicyThreadLocal.removePolicy();
     }
 }
