+++ extension/standard/src/main/java/cloud/piranha/extension/standard/StandardExtension.java
@@ -32,13 +32,13 @@ import cloud.piranha.core.api.WebApplicationExtensionContext;
 import cloud.piranha.extension.apache.fileupload.ApacheMultiPartExtension;
 import cloud.piranha.extension.annotationscan.AnnotationScanExtension;
 import cloud.piranha.extension.herring.HerringExtension;
-import cloud.piranha.extension.policy.PolicyExtension;
 import cloud.piranha.extension.security.servlet.ServletSecurityExtension;
 import cloud.piranha.extension.security.servlet.ServletSecurityManagerExtension;
 import cloud.piranha.extension.standard.async.StandardAsyncExtension;
 import cloud.piranha.extension.standard.localeencoding.StandardLocaleEncodingExtension;
 import cloud.piranha.extension.standard.logging.StandardLoggingExtension;
 import cloud.piranha.extension.standard.mimetype.StandardMimeTypeExtension;
+import cloud.piranha.extension.standard.policy.StandardPolicyExtension;
 import cloud.piranha.extension.standard.scinitializer.StandardServletContainerInitializerExtension;
 import cloud.piranha.extension.standard.servletannotations.StandardServletAnnotationsExtension;
 import cloud.piranha.extension.standard.tempdir.StandardTempDirExtension;
@@ -60,13 +60,13 @@ public class StandardExtension implements WebApplicationExtension {
         context.add(StandardLocaleEncodingExtension.class);
         context.add(StandardLoggingExtension.class);
         context.add(StandardMimeTypeExtension.class);
+        context.add(StandardPolicyExtension.class);
         context.add(StandardTempDirExtension.class);
         context.add(StandardWelcomeFileExtension.class);
         context.add(ServletSecurityManagerExtension.class);
         context.add(ApacheMultiPartExtension.class);
         context.add(WaspJspManagerExtension.class);
         context.add(HerringExtension.class);
-        context.add(PolicyExtension.class);
         context.add(AnnotationScanExtension.class);
         context.add(StandardWebXmlExtension.class);
         context.add(StandardServletAnnotationsExtension.class);
