+++ extension/standard/src/main/java/module-info.java
@@ -35,7 +35,6 @@
  * <ul>
  *  <li>Annotation scanning support</li>
  *  <li>JNDI support (Manorrock Herring)</li>
- *  <li>Policy support</li>
  *  <li>WaSP support</li>
  * </ul>
  */
@@ -45,12 +44,12 @@ module cloud.piranha.extension.standard {
     requires cloud.piranha.extension.apache.fileupload;
     requires cloud.piranha.extension.annotationscan;
     requires cloud.piranha.extension.herring;
-    requires cloud.piranha.extension.policy;
     requires cloud.piranha.extension.security.servlet;
     requires cloud.piranha.extension.standard.async;
     requires cloud.piranha.extension.standard.localeencoding;
     requires cloud.piranha.extension.standard.logging;
     requires cloud.piranha.extension.standard.mimetype;
+    requires cloud.piranha.extension.standard.policy;
     requires cloud.piranha.extension.standard.servletannotations;
     requires cloud.piranha.extension.standard.scinitializer;
     requires cloud.piranha.extension.standard.tempdir;
